Extract duplicate code into ResourceParsing module
